Add files via upload
This is about information retrieval tasks which involve finding the word frequencies in text document and finding the relationship between the frequency of words in text documents. The task is carried using python programming language in the file named frequencies.py and the used document is named wikitexts.txt. The output of the task is showed in the files named frequencies.png and frequencies.txt.